Extract media player hass URL handling to helper
