+++ homeassistant/components/roku/media_player.py
@@ -4,17 +4,15 @@ from __future__ import annotations
 import datetime as dt
 import logging
 from typing import Any
-from urllib.parse import quote
 
 import voluptuous as vol
-import yarl
 
 from homeassistant.components import media_source
-from homeassistant.components.http.auth import async_sign_path
 from homeassistant.components.media_player import (
     BrowseMedia,
     MediaPlayerDeviceClass,
     MediaPlayerEntity,
+    async_process_play_media_url,
 )
 from homeassistant.components.media_player.const import (
     ATTR_MEDIA_EXTRA,
@@ -47,7 +45,6 @@ from homeassistant.const import (
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_platform
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.network import get_url, is_hass_url
 
 from . import roku_exception_handler
 from .browse_media import async_browse_media
@@ -376,22 +373,8 @@ class RokuMediaPlayer(RokuEntity, MediaPlayerEntity):
             media_type = MEDIA_TYPE_URL
             media_id = sourced_media.url
 
-        # Sign and prefix with URL if playing a relative URL
-        if media_id[0] == "/" or is_hass_url(self.hass, media_id):
-            parsed = yarl.URL(media_id)
-
-            if parsed.query:
-                _LOGGER.debug("Not signing path for content with query param")
-            else:
-                media_id = async_sign_path(
-                    self.hass,
-                    quote(media_id),
-                    dt.timedelta(seconds=media_source.DEFAULT_EXPIRY_TIME),
-                )
-
-            # prepend external URL
-            if media_id[0] == "/":
-                media_id = f"{get_url(self.hass)}{media_id}"
+        # If media ID is a relative URL, we serve it from HA.
+        media_id = async_process_play_media_url(self.hass, media_id)
 
         if media_type not in PLAY_MEDIA_SUPPORTED_TYPES:
             _LOGGER.error(
